requirements.txt
Django>=5.0
python-dotenv
google-generativeai
